@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_translatedcontentarticle.
  *
- * (c) 2012-2022 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -16,16 +16,16 @@
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     Marc Reimann <reimann@mediendepot-ruhr.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2022 The MetaModels team.
+ * @copyright  2012-2024 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_translatedcontentarticle/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
-
 namespace MetaModels\AttributeTranslatedContentArticleBundle\Attribute;
 
+use Contao\ContentModel;
 use Contao\Controller;
-use MetaModels\AttributeTranslatedContentArticleBundle\Widgets\ContentArticleWidget;
+use Contao\Model\Collection;
 use MetaModels\Attribute\TranslatedReference;
 use MetaModels\ITranslatedMetaModel;
 
@@ -39,7 +39,7 @@ class TranslatedContentArticle extends TranslatedReference
      *
      * @var array
      */
-    private static $arrCallIds = [];
+    private static array $arrCallIds = [];
 
     /**
      * {@inheritdoc}
@@ -67,6 +67,7 @@ class TranslatedContentArticle extends TranslatedReference
     public function getFilterOptions($idList, $usedOnly, &$arrCount = null)
     {
         // Needed to fake implement BaseComplex.
+        return [];
     }
 
     /**
@@ -84,7 +85,6 @@ class TranslatedContentArticle extends TranslatedReference
     {
         // Needed to fake implement ITranslate.
     }
-
 
     /**
      * {@inheritdoc}
@@ -110,23 +110,23 @@ class TranslatedContentArticle extends TranslatedReference
      */
     public function getTranslatedDataFor($arrIds, $strLangCode)
     {
-        $strTable       = $this->getMetaModel()->getTableName();
-        $strColumn      = $this->getColName();
-        $model          = $this->getMetaModel();
-        $arrData        = [];
-        $contentArticle = new ContentArticleWidget();
-        $rootTable      = $contentArticle->getRootMetaModelTable($strTable);
+        $strTable  = $this->getMetaModel()->getTableName();
+        $strColumn = $this->getColName();
+        $model     = $this->getMetaModel();
+        $arrData   = [];
 
-        if ($this->getMetaModel() instanceof ITranslatedMetaModel) {
+        if ($model instanceof ITranslatedMetaModel) {
             $strLanguage         = $strLangCode;
             $strFallbackLanguage = $model->getMainLanguage();
         } else {
-            $strLanguage         = $model->isTranslated() ? $strLangCode : '';
+            /** @psalm-suppress DeprecatedMethod */
+            $strLanguage = $model->isTranslated() ? $strLangCode : '';
+            /** @psalm-suppress DeprecatedMethod */
             $strFallbackLanguage = $model->isTranslated() ? $model->getFallbackLanguage() : '';
         }
 
         foreach ($arrIds as $intId) {
-            // Continue if it's a recursive call
+            // Continue if it's a recursive call.
             $strCallId = $strTable . '_' . $strColumn . '_' . $strLanguage . '_' . $intId;
             if (isset(static::$arrCallIds[$strCallId])) {
                 $arrData[$intId]['value'] = \sprintf('RECURSION: %s', $strCallId);
@@ -134,58 +134,33 @@ class TranslatedContentArticle extends TranslatedReference
             }
             static::$arrCallIds[$strCallId] = true;
 
-            // Generate list for backend.
-            if (TL_MODE == 'BE') {
-                $elements = $contentArticle->getContentTypesByRecordId($intId, $rootTable, $strColumn, $strLanguage);
-                $content  = '';
-                if (count($elements)) {
-                    $content .= '<ul class="elements_container">';
-                    foreach ((array) $elements as $element) {
-                        // @codingStandardsIgnoreStart - one line template.
-                        $content .= \sprintf(
-                            '<li><div class="cte_type%s"><img src="system/themes/flexible/icons/%s.svg" width="16" height="16"> %s</div></li>',
-                            $element['isInvisible'] ? ' unpublished' : ' published',
-                            $element['isInvisible'] ? 'invisible' : 'visible',
-                            $element['name']
-                        );
-                        // @codingStandardsIgnoreEnd
-                    }
-                    $content .= '</ul>';
-                }
+            $objContent         = ContentModel::findPublishedByPidAndTable((int) $intId, $strTable);
+            $arrContent         = [];
+            $arrContentFallback = [];
 
-                if (!empty($content)) {
-                    $arrData[$intId]['value'] = [$content];
-                } else {
-                    $arrData[$intId]['value'] = [];
+            if ($objContent !== null) {
+                assert($objContent instanceof Collection);
+                while ($objContent->next()) {
+                    /** @psalm-suppress UndefinedMagicPropertyFetch */
+                    if ($objContent->mm_slot === $strColumn && $objContent->mm_lang === $strLanguage) {
+                        $arrContent[] = Controller::getContentElement($objContent->current());
+                    } elseif (
+                        /** @psalm-suppress UndefinedMagicPropertyFetch */
+                        $objContent->mm_slot === $strColumn
+                        && $strLanguage !== $strFallbackLanguage
+                        && $objContent->mm_lang === $strFallbackLanguage
+                    ) {
+                        $arrContentFallback[] = Controller::getContentElement($objContent->current());
+                    }
                 }
             }
 
-            // Generate output for frontend.
-            if (TL_MODE == 'FE') {
-                $objContent         = \ContentModel::findPublishedByPidAndTable($intId, $strTable);
-                $arrContent         = [];
-                $arrContentFallback = [];
-
-                if ($objContent !== null) {
-                    while ($objContent->next()) {
-                        if ($objContent->mm_slot == $strColumn && $objContent->mm_lang == $strLanguage) {
-                            $arrContent[] = Controller::getContentElement($objContent->current());
-                        } elseif ($objContent->mm_slot == $strColumn
-                                  && $strLanguage != $strFallbackLanguage
-                                  && $objContent->mm_lang == $strFallbackLanguage
-                        ) {
-                            $arrContentFallback[] = Controller::getContentElement($objContent->current());
-                        }
-                    }
-                }
-
-                if (!empty($arrContent)) {
-                    $arrData[$intId]['value'] = $arrContent;
-                } elseif (!empty($arrContentFallback)) {
-                    $arrData[$intId]['value'] = $arrContentFallback;
-                } else {
-                    $arrData[$intId]['value'] = [];
-                }
+            if (!empty($arrContent)) {
+                $arrData[$intId]['value'] = $arrContent;
+            } elseif (!empty($arrContentFallback)) {
+                $arrData[$intId]['value'] = $arrContentFallback;
+            } else {
+                $arrData[$intId]['value'] = [];
             }
 
             unset(static::$arrCallIds[$strCallId]);
