@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_translatedcontentarticle.
  *
- * (c) 2012-2021 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,13 +15,10 @@
  * @author     Andreas Dziemba <adziemba@web.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
- * @copyright  2012-2021 The MetaModels team.
+ * @copyright  2012-2024 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_translatedcontentarticle/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
-
-use Contao\Input;
-use MetaModels\AttributeTranslatedContentArticleBundle\Table\ArticleContent;
 
 $GLOBALS['TL_DCA']['tl_content']['fields']['mm_slot'] = [
     'sql' => "varchar(255) NOT NULL default ''",
@@ -30,48 +27,3 @@ $GLOBALS['TL_DCA']['tl_content']['fields']['mm_slot'] = [
 $GLOBALS['TL_DCA']['tl_content']['fields']['mm_lang'] = [
     'sql' => "varchar(5) NOT NULL default ''",
 ];
-
-$strModule      = Input::get('do');
-$strTable       = Input::get('table');
-$strLangSupport = Input::get('langSupport');
-
-// Change TL_Content for the article popup
-if (\substr($strModule, 0, 10) == 'metamodel_' && $strTable == 'tl_content' && $strLangSupport == '1') {
-    $GLOBALS['TL_DCA']['tl_content']['config']['ptable']                                =
-        Input::get('ptable');
-    $GLOBALS['TL_DCA']['tl_content']['config']['onsubmit_callback'][]                   =
-        [
-            ArticleContent::class,
-            'save'
-        ];
-    $GLOBALS['TL_DCA']['tl_content']['config']['onload_callback'][]                     =
-        [
-            ArticleContent::class,
-            'checkPermission'
-        ];
-    $GLOBALS['TL_DCA']['tl_content']['config']['oncopy_callback'][]                     =
-        [
-            ArticleContent::class,
-            'updateCopyData'
-        ];
-    $GLOBALS['TL_DCA']['tl_content']['config']['oncut_callback'][]                      =
-        [
-            ArticleContent::class,
-            'updateCutData'
-        ];
-    $GLOBALS['TL_DCA']['tl_content']['list']['operations']['toggle']['button_callback'] =
-        [
-            ArticleContent::class,
-            'toggleIcon'
-        ];
-    $GLOBALS['TL_DCA']['tl_content']['list']['sorting']['filter'][]                     =
-        [
-            'mm_slot=?',
-            Input::get('slot')
-        ];
-    $GLOBALS['TL_DCA']['tl_content']['list']['sorting']['filter'][]                     =
-        [
-            'mm_lang=?',
-            Input::get('lang')
-        ];
-}
