@@ -1,11 +1,9 @@
 <?php
 
-declare(strict_types=1);
-
 /**
- * This file is part of MetaModels/attribute_contentarticle.
+ * This file is part of MetaModels/attribute_translatedcontentarticle.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,15 +11,20 @@ declare(strict_types=1);
  * This project is provided in good faith and hope to be usable by anyone.
  *
  * @package    MetaModels
- * @subpackage AttributeContentArticle
+ * @subpackage AttributeTranslatedContentArticle
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2012-2024 The MetaModels team.
+ * @license    https://github.com/MetaModels/attribute_translatedcontentarticle/blob/master/LICENSE LGPL-3.0-or-later
+ * @filesource
  */
+
+declare(strict_types=1);
 
 namespace ContaoManager;
 
 use Contao\ManagerPlugin\Bundle\Config\ConfigInterface;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
-use MetaModels\AttributeContentArticleBundle\MetaModelsAttributeContentArticleBundle;
 use MetaModels\AttributeTranslatedContentArticleBundle\ContaoManager\Plugin;
 use MetaModels\AttributeTranslatedContentArticleBundle\MetaModelsAttributeTranslatedContentArticleBundle;
 use MetaModels\CoreBundle\MetaModelsCoreBundle;
@@ -44,7 +47,7 @@ class PluginTest extends TestCase
         $config
             ->expects(self::once())
             ->method('getLoadAfter')
-            ->willReturn([MetaModelsAttributeContentArticleBundle::class, MetaModelsCoreBundle::class]);
+            ->willReturn([MetaModelsCoreBundle::class]);
         $config
             ->expects(self::once())
             ->method('getReplace')
