@@ -1,11 +1,9 @@
 <?php
 
-declare(strict_types=1);
-
 /**
- * This file is part of MetaModels/attribute_contentarticle.
+ * This file is part of MetaModels/attribute_translatedcontentarticle.
  *
- * (c) 2012-2022 The MetaModels team.
+ * (c) 2012-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,30 +11,44 @@ declare(strict_types=1);
  * This project is provided in good faith and hope to be usable by anyone.
  *
  * @package    MetaModels
- * @subpackage AttributeContentArticle
+ * @subpackage AttributeTranslatedContentArticle
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     Sven Baumann <baumann.sv@gmail.com>
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2012-2025 The MetaModels team.
+ * @license    https://github.com/MetaModels/attribute_translatedcontentarticle/blob/master/LICENSE LGPL-3.0-or-later
+ * @filesource
  */
+
+declare(strict_types=1);
 
 namespace DependencyInjection;
 
+// phpcs:disable
+use MetaModels\AttributeTranslatedContentArticleBundle\Controller\Backend\MetaModelController;
 use MetaModels\AttributeTranslatedContentArticleBundle\Attribute\AttributeTypeFactory;
 use MetaModels\AttributeTranslatedContentArticleBundle\EventListener\BackendEventListener;
 use MetaModels\AttributeTranslatedContentArticleBundle\EventListener\GetOptionsListener;
-use MetaModels\AttributeTranslatedContentArticleBundle\EventListener\InitializeListener;
 use MetaModels\AttributeTranslatedContentArticleBundle\DependencyInjection\MetaModelsAttributeTranslatedContentArticleExtension;
-use MetaModels\AttributeTranslatedContentArticleBundle\Table\ArticleContent;
+use MetaModels\AttributeTranslatedContentArticleBundle\MetaModelsAttributeTranslatedContentArticleBundle;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+// phpcs:enable
 
 /**
+ * phpcs:disable
  * @covers \MetaModels\AttributeTranslatedContentArticleBundle\DependencyInjection\MetaModelsAttributeContentArticleExtension
+ * phpcs:enable
+ *
+ * @SuppressWarnings(PHPMD.LongClassName)
  */
 class MetaModelsAttributeTranslatedContentArticleExtensionTest extends TestCase
 {
     public function testLoad(): void
     {
         $container = new ContainerBuilder();
+        $container->setParameter('kernel.bundles', [MetaModelsAttributeTranslatedContentArticleBundle::class]);
+
         $extension = new MetaModelsAttributeTranslatedContentArticleExtension();
         $extension->load([], $container);
 
@@ -44,9 +56,8 @@ class MetaModelsAttributeTranslatedContentArticleExtensionTest extends TestCase
             'service_container',
             BackendEventListener::class,
             GetOptionsListener::class,
-            InitializeListener::class,
             AttributeTypeFactory::class,
-            ArticleContent::class
+            MetaModelController::class,
         ];
 
         self::assertCount(count($expectedDefinitions), $container->getDefinitions());
